pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dr-odir-classifier"
version = "0.1.0"
description = "Reproducible ODIR diabetic retinopathy classification benchmark."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.2.0",
    "Pillow>=9.5.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = ["black", "ruff"]

